+++ shuffle.h
@@ -1,6 +1,6 @@
 #ifndef HAS_SHUFFLE_H
 #define HAS_SHUFFLE_H
 
-void shuffleptr(void* *array, size_t n);
+void shuffleptr(void* *array, unsigned int n);
 
 #endif
